Clarify MAVLink over IP
MAVLink over IP is confusing in how it sets up the client and server. GCS in effect acts as a (network) server: it listens on a specific port; AP acts as a client, sending packets to a known address/port combination. This is the opposite of how it sets up a serial port, where anyone who can establish a serial connection can start listening to MAVLink from AP. For a reader who expects GCS to be the client that connects to an AP server on a known address, the text I'm editing would deepen the confusion.

My copy isn't perfect, and if @Hwurzburg or anyone else has a better way to express this idea, I support that. I'm simply trying to clarify the assumptions about which side is listening and which side need to know the other's IP address.